+++ service/testing.go
@@ -39,6 +39,9 @@ func (s *Tier1Service) TestBlocks(ctx context.Context, isSubRequest bool, reques
 	if err != nil {
 		return stream.NewErrInvalidArg(err.Error())
 	}
+	s.checkPendingShutdown = func() bool {
+		return false
+	}
 
 	return s.blocks(ctx, request, execGraph, respFunc)
 }