+++ service/tier1.go
@@ -81,6 +81,7 @@ type Tier1Service struct {
 	activeRequestsHardLimit int
 	tier2RequestParameters  reqctx.Tier2RequestParameters
 	globalRequestPool       *GlobalRequestPool
+	checkPendingShutdown    func() bool
 }
 
 func getBlockTypeFromStreamFactory(sf *StreamFactory) (string, error) {
@@ -145,6 +146,7 @@ func NewTier1(
 	activeRequestsHardLimit int,
 	sharedCacheSize uint64,
 	globalRequestPool *GlobalRequestPool,
+	checkPendingShutdown func() bool,
 	opts ...Option,
 ) (*Tier1Service, error) {
 
@@ -193,6 +195,7 @@ func NewTier1(
 		activeRequestsSoftLimit: activeRequestsSoftLimit,
 		activeRequestsHardLimit: activeRequestsHardLimit,
 		globalRequestPool:       globalRequestPool,
+		checkPendingShutdown:    checkPendingShutdown,
 	}
 
 	go func() {
@@ -243,6 +246,11 @@ func (s *Tier1Service) Blocks(
 			metrics.Tier1RejectedRequestCounter.Inc(reason)
 		}
 	}()
+	if s.checkPendingShutdown() {
+		time.Sleep(1 * time.Second) // prevent burst reconnections to this instance since we're shutting down
+		serverErr = connect.NewError(connect.CodeUnavailable, errShuttingDown)
+		return
+	}
 
 	// We keep `err` here as the unaltered error from `blocks` call, this is used in the EndSpan to record the full error
 	// and not only the `grpcError` one which is a subset view of the full `err`.
@@ -611,6 +619,7 @@ func (s *Tier1Service) blocks(ctx context.Context, request *pbsubstreamsrpc.Requ
 
 	pipe := pipeline.New(
 		ctx,
+		true,
 		execGraph,
 		stores,
 		nil,
@@ -621,6 +630,7 @@ func (s *Tier1Service) blocks(ctx context.Context, request *pbsubstreamsrpc.Requ
 		s.runtimeConfig.WorkerPoolFactory,
 		respFunc,
 		s.blockExecutionTimeout,
+		s.checkPendingShutdown,
 		opts...,
 	)
 
@@ -847,6 +857,11 @@ func toConnectError(ctx context.Context, err error) error {
 			}
 		}
 		return connect.NewError(connect.CodeCanceled, err)
+	}
+
+	// special case for "QuickSave" on shutdown
+	if err == pipeline.ErrShuttingDown {
+		return connect.NewError(connect.CodeUnavailable, err)
 	}
 
 	// context deadline exceeded